spring-boot-project/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/scheduling/SchedulingAutoConfiguration.java
/*
 * Copyright 2012-2018 the original author or authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package org.springframework.boot.autoconfigure.scheduling;

import org.springframework.boot.autoconfigure.AutoConfigureAfter;
import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
import org.springframework.boot.autoconfigure.condition.AnyNestedCondition;
import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
import org.springframework.boot.autoconfigure.task.TaskExecutorAutoConfiguration;
import org.springframework.boot.task.TaskExecutorBuilder;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Conditional;
import org.springframework.context.annotation.Configuration;
import org.springframework.core.task.TaskExecutor;
import org.springframework.scheduling.annotation.AsyncConfigurer;
import org.springframework.scheduling.annotation.EnableAsync;
import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
import org.springframework.scheduling.config.TaskManagementConfigUtils;

/**
 * {@link EnableAutoConfiguration Auto-configuration} for scheduling.
 *
 * @author Stephane Nicoll
 * @since 2.1.0
 * @see EnableAsync
 */
@Configuration
@ConditionalOnMissingBean(TaskExecutor.class)
@AutoConfigureAfter(TaskExecutorAutoConfiguration.class)
public class SchedulingAutoConfiguration {

	@Configuration
	@Conditional(EnableAsyncAvailableCondition.class)
	@ConditionalOnMissingBean(AsyncConfigurer.class)
	static class AsyncConfiguration {

		@Bean
		public ThreadPoolTaskExecutor taskExecutor(TaskExecutorBuilder builder) {
			return builder.build();
		}

	}

	static class EnableAsyncAvailableCondition extends AnyNestedCondition {

		EnableAsyncAvailableCondition() {
			super(ConfigurationPhase.REGISTER_BEAN);
		}

		@ConditionalOnBean(name = TaskManagementConfigUtils.ASYNC_ANNOTATION_PROCESSOR_BEAN_NAME)
		static class ProxyMode {

		}

		@ConditionalOnBean(name = TaskManagementConfigUtils.ASYNC_EXECUTION_ASPECT_BEAN_NAME)
		static class AspectJMode {

		}

	}

}
